+++ app/src/main/java/com/internship/ipda3/semicolon/bloodbank/model/general/governorates/GavernoratesDatum.java
@@ -9,7 +9,7 @@ public class GavernoratesDatum {
     @SerializedName("created_at")
     private Object mCreatedAt;
     @SerializedName("id")
-    private Long mId;
+    private Integer mId;
     @SerializedName("name")
     private String mName;
     @SerializedName("updated_at")
@@ -23,11 +23,11 @@ public class GavernoratesDatum {
         mCreatedAt = createdAt;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return mId;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         mId = id;
     }
